# backend/tests/data/openapi.json
{
  "components": {
    "schemas": {
...
      "CharacterCreate": {
        "properties": {
          "bleed_marks": {
            "default": 0,
            "title": "Bleed Marks",
            "type": "integer"
          },
          "body_marks": {
            "default": 0,
            "title": "Body Marks",
            "type": "integer"
          },
          "brain_marks": {
            "default": 0,
            "title": "Brain Marks",
            "type": "integer"
          },
          "catalyst": {
            "default": "",
            "title": "Catalyst",
            "type": "string"
          },
          "control": {
            "default": 0,
            "title": "Control",
            "type": "integer"
          },
          "cunning_current": {
            "default": 1,
            "title": "Cunning Current",
            "type": "integer"
          },
          "cunning_max": {
            "default": 1,
            "title": "Cunning Max",
            "type": "integer"
          },
          "cunning_resistance_spent": {
            "default": 0,
            "title": "Cunning Resistance Spent",
            "type": "integer"
          },
          "gear": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Gear",
            "type": "array"
          },
          "gilded_control": {
            "default": false,
            "title": "Gilded Control",
            "type": "boolean"
          },
          "gilded_hide": {
            "default": false,
            "title": "Gilded Hide",
            "type": "boolean"
          },
          "gilded_move": {
            "default": false,
            "title": "Gilded Move",
            "type": "boolean"
          },
          "gilded_read": {
            "default": false,
            "title": "Gilded Read",
            "type": "boolean"
          },
          "gilded_sense": {
            "default": false,
            "title": "Gilded Sense",
            "type": "boolean"
          },
          "gilded_sneak": {
            "default": false,
            "title": "Gilded Sneak",
            "type": "boolean"
          },
          "gilded_strike": {
            "default": false,
            "title": "Gilded Strike",
            "type": "boolean"
          },
          "gilded_survey": {
            "default": false,
            "title": "Gilded Survey",
            "type": "boolean"
          },
          "gilded_sway": {
            "default": false,
            "title": "Gilded Sway",
            "type": "boolean"
          },
          "hide": {
            "default": 0,
            "title": "Hide",
            "type": "integer"
          },
          "incapacitated": {
            "default": false,
            "title": "Incapacitated",
            "type": "boolean"
          },
          "intuition_current": {
            "default": 1,
            "title": "Intuition Current",
            "type": "integer"
          },
          "intuition_max": {
            "default": 1,
            "title": "Intuition Max",
            "type": "integer"
          },
          "intuition_resistance_spent": {
            "default": 0,
            "title": "Intuition Resistance Spent",
            "type": "integer"
          },
          "move": {
            "default": 0,
            "title": "Move",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "nerve_current": {
            "default": 1,
            "title": "Nerve Current",
            "type": "integer"
          },
          "nerve_max": {
            "default": 1,
            "title": "Nerve Max",
            "type": "integer"
          },
          "nerve_resistance_spent": {
            "default": 0,
            "title": "Nerve Resistance Spent",
            "type": "integer"
          },
          "profile_pic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Pic"
          },
          "pronouns": {
            "default": "Unlisted",
            "title": "Pronouns",
            "type": "string"
          },
          "question": {
            "default": "",
            "title": "Question",
            "type": "string"
          },
          "read": {
            "default": 0,
            "title": "Read",
            "type": "integer"
          },
          "role": {
            "default": "",
            "title": "Role",
            "type": "string"
          },
          "role_ability": {
            "default": "None",
            "title": "Role Ability",
            "type": "string"
          },
          "scars_count": {
            "default": 0,
            "title": "Scars Count",
            "type": "integer"
          },
          "scars_list": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Scars List",
            "type": "array"
          },
          "sense": {
            "default": 0,
            "title": "Sense",
            "type": "integer"
          },
          "sneak": {
            "default": 0,
            "title": "Sneak",
            "type": "integer"
          },
          "specialty": {
            "default": "",
            "title": "Specialty",
            "type": "string"
          },
          "specialty_ability": {
            "default": "None",
            "title": "Specialty Ability",
            "type": "string"
          },
          "strike": {
            "default": 0,
            "title": "Strike",
            "type": "integer"
          },
          "style": {
            "default": "",
            "title": "Style",
            "type": "string"
          },
          "survey": {
            "default": 0,
            "title": "Survey",
            "type": "integer"
          },
          "sway": {
            "default": 0,
            "title": "Sway",
            "type": "integer"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "name"
        ],
        "title": "CharacterCreate",
        "type": "object"
      },
      "CharacterResponse": {
        "properties": {
          "bleed_marks": {
            "default": 0,
            "title": "Bleed Marks",
            "type": "integer"
          },
          "body_marks": {
            "default": 0,
            "title": "Body Marks",
            "type": "integer"
          },
          "brain_marks": {
            "default": 0,
            "title": "Brain Marks",
            "type": "integer"
          },
          "catalyst": {
            "default": "",
            "title": "Catalyst",
            "type": "string"
          },
          "circle_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Circle Id"
          },
          "control": {
            "default": 0,
            "title": "Control",
            "type": "integer"
          },
          "cunning_current": {
            "default": 1,
            "title": "Cunning Current",
            "type": "integer"
          },
          "cunning_max": {
            "default": 1,
            "title": "Cunning Max",
            "type": "integer"
          },
          "cunning_resistance_spent": {
            "default": 0,
            "title": "Cunning Resistance Spent",
            "type": "integer"
          },
          "gear": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Gear",
            "type": "array"
          },
          "gilded_control": {
            "default": false,
            "title": "Gilded Control",
            "type": "boolean"
          },
          "gilded_hide": {
            "default": false,
            "title": "Gilded Hide",
            "type": "boolean"
          },
          "gilded_move": {
            "default": false,
            "title": "Gilded Move",
            "type": "boolean"
          },
          "gilded_read": {
            "default": false,
            "title": "Gilded Read",
            "type": "boolean"
          },
          "gilded_sense": {
            "default": false,
            "title": "Gilded Sense",
            "type": "boolean"
          },
          "gilded_sneak": {
            "default": false,
            "title": "Gilded Sneak",
            "type": "boolean"
          },
          "gilded_strike": {
            "default": false,
            "title": "Gilded Strike",
            "type": "boolean"
          },
          "gilded_survey": {
            "default": false,
            "title": "Gilded Survey",
            "type": "boolean"
          },
          "gilded_sway": {
            "default": false,
            "title": "Gilded Sway",
            "type": "boolean"
          },
          "hide": {
            "default": 0,
            "title": "Hide",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "incapacitated": {
            "default": false,
            "title": "Incapacitated",
            "type": "boolean"
          },
          "ink_color": {
            "default": "",
            "title": "Ink Color",
            "type": "string"
          },
          "intuition_current": {
            "default": 1,
            "title": "Intuition Current",
            "type": "integer"
          },
          "intuition_max": {
            "default": 1,
            "title": "Intuition Max",
            "type": "integer"
          },
          "intuition_resistance_spent": {
            "default": 0,
            "title": "Intuition Resistance Spent",
            "type": "integer"
          },
          "move": {
            "default": 0,
            "title": "Move",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "nerve_current": {
            "default": 1,
            "title": "Nerve Current",
            "type": "integer"
          },
          "nerve_max": {
            "default": 1,
            "title": "Nerve Max",
            "type": "integer"
          },
          "nerve_resistance_spent": {
            "default": 0,
            "title": "Nerve Resistance Spent",
            "type": "integer"
          },
          "pen_font": {
            "default": "Caveat",
            "title": "Pen Font",
            "type": "string"
          },
          "profile_pic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Pic"
          },
          "pronouns": {
            "default": "Unlisted",
            "title": "Pronouns",
            "type": "string"
          },
          "question": {
            "default": "",
            "title": "Question",
            "type": "string"
          },
          "read": {
            "default": 0,
            "title": "Read",
            "type": "integer"
          },
          "role": {
            "default": "",
            "title": "Role",
            "type": "string"
          },
          "role_ability": {
            "default": "None",
            "title": "Role Ability",
            "type": "string"
          },
          "scars_count": {
            "default": 0,
            "title": "Scars Count",
            "type": "integer"
          },
          "scars_list": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Scars List",
            "type": "array"
          },
          "sense": {
            "default": 0,
            "title": "Sense",
            "type": "integer"
          },
          "sneak": {
            "default": 0,
            "title": "Sneak",
            "type": "integer"
          },
          "specialty": {
            "default": "",
            "title": "Specialty",
            "type": "string"
          },
          "specialty_ability": {
            "default": "None",
            "title": "Specialty Ability",
            "type": "string"
          },
          "status": {
            "default": "unaffiliated",
            "title": "Status",
            "type": "string"
          },
          "strike": {
            "default": 0,
            "title": "Strike",
            "type": "integer"
          },
          "style": {
            "default": "",
            "title": "Style",
            "type": "string"
          },
          "survey": {
            "default": 0,
            "title": "Survey",
            "type": "integer"
          },
          "sway": {
            "default": 0,
            "title": "Sway",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "id"
        ],
        "title": "CharacterResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PortraitUpdate": {
        "properties": {
          "profile_pic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Pic"
          }
        },
        "required": [
          "profile_pic"
        ],
        "title": "PortraitUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/investigators/forge": {
      "post": {
        "operationId": "forge_investigator_api_investigators_forge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Forge Investigator"
      }
    },
    "/api/investigators/{investigator_id}": {
      "delete": {
        "description": "Deletes the caller's own character, which must be on no campaign's roster (409\nwhile it is active or pending in one). Nobody else may, the GM of its campaign\nincluded. A soft delete: the owner can undo it for a short while (POST\n.../restore) and an admin can restore it later (docs/refactor/DELETION.md). A\nsocket still open on the character's channel gets character_deleted and is closed\nwith 4404.",
        "operationId": "delete_investigator_api_investigators__investigator_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "investigator_id",
            "required": true,
            "schema": {
              "title": "Investigator Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Investigator"
      },
      "get": {
        "operationId": "get_investigator_api_investigators__investigator_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "investigator_id",
            "required": true,
            "schema": {
              "title": "Investigator Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Investigator"
      }
    },
    "/api/investigators/{investigator_id}/portrait": {
      "put": {
        "description": "Sets the character's portrait to a picture data URL, or clears it (profile_pic\nnull). Allowed for the owner, and for the GM of the character's campaign while it\nis active or pending, a limited number of times per user (vtt/portraits.py).\nAnswers with the character as the WebSocket sends it.",
        "operationId": "set_portrait_api_investigators__investigator_id__portrait_put",
        "parameters": [
          {
            "in": "path",
            "name": "investigator_id",
            "required": true,
            "schema": {
              "title": "Investigator Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortraitUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Portrait"
      }
    },
    "/api/investigators/{investigator_id}/restore": {
      "post": {
        "description": "Undoes the caller's delete of their own character, within deletion.UNDO_SECONDS\n(409 after that). A character that is not deleted, or not the caller's, is 404.",
        "operationId": "restore_investigator_api_investigators__investigator_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "investigator_id",
            "required": true,
            "schema": {
              "title": "Investigator Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Investigator"
      }
    },
...
  }
}